refactor: extract isUuid into shared util, rename id→idOrName in get_* tools
- Create src/utils/uuid.ts with canonical isUuid() helper
- models/interaction.ts: re-export isUuid from shared util (remove inline def)
- models/agent.ts: import isUuid from utils/uuid instead of models/interaction
- archestra-mcp-server/agents.ts:
  - import isUuid from utils/uuid
  - rename id→idOrName in get_agent/get_llm_proxy/get_mcp_gateway tool schemas
  - update handler variable and logging field accordingly
- routes/chatops.ts: remove local UUID_REGEX+isValidUUID, import isUuid from utils/uuid
- docs: update get_agent/get_llm_proxy/get_mcp_gateway parameter tables (codegen output)
