+++ platform/backend/src/utils/uuid.ts
@@ -1,0 +1,9 @@
+const UUID_REGEX =
+  /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+
+/**
+ * Check if a string is a valid UUID format.
+ */
+export function isUuid(str: string): boolean {
+  return UUID_REGEX.test(str);
+}
